Update README.md
This was my first project in the Data Science field.
Being that, I still did not have strong Python basics and it had not been structured in a proper way (e.g. functions are defined using local variables, no classes are used etcetera). Furthermore italian naming was used to define variables and functions.

Please reference the Poster and the paper for further details about the procedure.